Enable server actions in app router
